refine reshard
