+++ paddle/phi/api/lib/tensor_copy.cc
@@ -26,7 +26,7 @@ limitations under the License. */
 #include "paddle/phi/infermeta/unary.h"
 #ifdef PADDLE_WITH_DISTRIBUTE
 #include "paddle/phi/api/lib/data_transform.h"
-#include "paddle/phi/core/distributed/auto_parallel/reshard_utils.h"
+#include "paddle/phi/core/distributed/auto_parallel/reshard/reshard_utils.h"
 #include "paddle/phi/infermeta/spmd_rules/rules.h"
 #endif
 namespace paddle {
