+++ paddle/phi/core/distributed/auto_parallel/reshard/CMakeLists.txt
@@ -1,0 +1,12 @@
+collect_srcs(
+  core_srcs
+  SRCS
+  reshard_utils.cc
+  reshard_function.cc
+  r_to_s_reshard_function.cc
+  s_to_r_reshard_function.cc
+  r_to_p_reshard_function.cc
+  p_to_r_reshard_function.cc
+  s_to_s_reshard_function.cc
+  nd_mesh_reshard_function.cc
+  same_status_reshard_function.cc)
