+++ paddle/phi/core/distributed/auto_parallel/reshard/nd_mesh_reshard_function.h
@@ -14,16 +14,13 @@
 
 #pragma once
 
-#include "paddle/phi/core/distributed/auto_parallel/reshard_function.h"
+#include "paddle/phi/core/distributed/auto_parallel/reshard/reshard_function.h"
 
 namespace phi {
 namespace distributed {
 
-class PToRReshardFunction final : public ReshardFunction {
+class SameNdMeshReshardFunction final : public ReshardFunction {
  public:
-  PToRReshardFunction() = default;
-  ~PToRReshardFunction() = default;
-
   bool IsSuitable(const DistTensor& in,
                   const TensorDistAttr& out_dist_attr) override;
