+++ paddle/phi/core/distributed/auto_parallel/reshard/p_to_r_reshard_function.h
@@ -14,15 +14,15 @@
 
 #pragma once
 
-#include "paddle/phi/core/distributed/auto_parallel/reshard_function.h"
+#include "paddle/phi/core/distributed/auto_parallel/reshard/reshard_function.h"
 
 namespace phi {
 namespace distributed {
 
-class SToSReshardFunction final : public ReshardFunction {
+class PToRReshardFunction final : public ReshardFunction {
  public:
-  SToSReshardFunction() = default;
-  ~SToSReshardFunction() = default;
+  PToRReshardFunction() = default;
+  ~PToRReshardFunction() = default;
 
   bool IsSuitable(const DistTensor& in,
                   const TensorDistAttr& out_dist_attr) override;
